move getPageSize
